mm/slab: fix cpu on/off handling
When cpu off, we flush all cpu cached objects to it's own slab. 
free_block() is used for this purpose and it's role is just to flush
objects from array_cache to proper slab.  It doesn't adjust array_cache's
internal fields so we should manually reset them to proper value.  Without
this fix, we maintain free objects duplicately, one is in cpu cache, and,
the other one is in the slab.  So system would be broken.

Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Reported-by: Jeremiah Mahler <jmmahler@gmail.com>
Cc: Christoph Lameter <cl@linux.com>
Cc: Pekka Enberg <penberg@kernel.org>
Cc: David Rientjes <rientjes@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>