cpuset: simplify cpuset_node_allowed API
Current cpuset API for checking if a zone/node is allowed to allocate from
looks rather awkward.  We have hardwall and softwall versions of
cpuset_node_allowed with the softwall version doing literally the same as
the hardwall version if __GFP_HARDWALL is passed to it in gfp flags.  If
it isn't, the softwall version may check the given node against the
enclosing hardwall cpuset, which it needs to take the callback lock to do.

Such a distinction was introduced by commit 02a0e53d8227 ("cpuset: rework
cpuset_zone_allowed api").  Before, we had the only version with the
__GFP_HARDWALL flag determining its behavior.  The purpose of the commit
was to avoid sleep-in-atomic bugs when someone would mistakenly call the
function without the __GFP_HARDWALL flag for an atomic allocation.  The
suffixes introduced were intended to make the callers think before using
the function.

However, since the callback lock was converted from mutex to spinlock by
the previous patch, the softwall check function cannot sleep, and these
precautions are no longer necessary.

So let's simplify the API back to the single check.

Signed-off-by: Vladimir Davydov <vdavydov@parallels.com>
Suggested-by: David Rientjes <rientjes@google.com>
Acked-by: Christoph Lameter <cl@linux.com>
Acked-by: Zefan Li <lizefan@huawei.com>
Cc: Pekka Enberg <penberg@kernel.org>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: Tejun Heo <tj@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>